+++ src/pykakasi/__init__.py
@@ -1,10 +1,7 @@
-from .a2 import a2
-from .h2 import H2
-from .j2 import J2
-from .k2 import K2
-from .kakasi import kakasi
-from .sym2 import sym2
-from .wakati import wakati
+from .kakasi import kakasi, wakati
+from .kana import H2, K2
+from .kanji import J2
+from .roman import a2, sym2
 
 __copyright__ = 'Copyright (C) 2010-2019 Hiroshi Miura'
 __version__ = '0.96'
