+++ src/pykakasi/kanji.py
@@ -26,9 +26,13 @@
 # */
 
 import re
+import threading
 
-from .itaiji import itaiji
-from .kanwa import kanwa
+from klepto.archives import file_archive
+from pkg_resources import resource_filename
+
+from .kana import H2
+from .properties import Configurations
 
 
 class J2(object):
@@ -50,7 +54,6 @@ class J2(object):
         if mode == "H":
             self.convert = self.convert_H
         elif mode in ("a", "K"):
-            from .h2 import H2
             self._hconv = H2(mode, method)
             self.convert = self.convert_nonH
         else:
@@ -120,3 +123,74 @@ class J2(object):
 
     def convert_noop(self, text):
         return text[0], 1
+
+
+class itaiji (object):
+
+    # this class is Borg/Singleton
+    _shared_state = {
+        '_itaijidict': None,
+        '_itaijidict_len': 0,
+        '_lock': threading.Lock()
+    }
+
+    def __new__(cls, *p, **k):
+        self = object.__new__(cls, *p, **k)
+        self.__dict__ = cls._shared_state
+        return self
+
+    def __init__(self):
+        if self._itaijidict is None:
+            with self._lock:
+                if self._itaijidict is None:
+                    itaijipath = resource_filename(__name__, Configurations.jisyo_itaiji)
+                    self._itaijidict = file_archive(itaijipath, {}, serialized=True)
+                    self._itaijidict.load()
+                    self._itaijidict_len = self._itaijidict['_max_key_len_']
+
+    def haskey(self, key):
+        return key in self._itaijidict
+
+    def lookup(self, key):
+        return self._itaijidict[key]
+
+    def convert(self, text):
+        r = []
+        for c in text:
+            if self.haskey(c):
+                r.append(c)
+        for c in r:
+            text = re.sub(c, self.lookup(c), text)
+        return text
+
+
+# This class is Borg/Singleton
+# It provides same results becase lookup from a static dictionary.
+# There is no state rather dictionary dbm.
+class kanwa (object):
+    _shared_state = {
+        '_lock': threading.Lock(),
+        '_jisyo_table': None
+    }
+
+    # Note: there is no invalidate mechanism for _jisyo_table data.
+    #       It can lead a large memory consumption in long live process.
+    #       maximum memory consumption will be several megabytes which is
+    #       a size of a dictionary.
+
+    def __new__(cls, *p, **k):
+        self = object.__new__(cls, *p, **k)
+        self.__dict__ = cls._shared_state
+        return self
+
+    def __init__(self):
+        if self._jisyo_table is None:
+            with self._lock:
+                if self._jisyo_table is None:
+                    dictpath = resource_filename(__name__, Configurations.jisyo_kanwa)
+                    self._jisyo_table = file_archive(dictpath, {}, serialized=True)
+                    self._jisyo_table.load()
+
+    def load(self, char):
+        key = "%04x" % ord(char)
+        return self._jisyo_table.get(key, None)
