+++ src/pykakasi/properties.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+#  properties.py
+#
+# Copyright 2019 Hiroshi Miura <miurahr@linux.com>
+
+
+class Configurations:
+    jisyo_hepburn_hira = 'hepburnhira3.db'
+    jisyo_passport_hira = 'passporthira3.db'
+    jisyo_kunrei_hira = 'kunreihira3.db'
+    jisyo_itaiji = 'itaijidict3.db'
+    jisyo_kanwa = 'kanwadict4.db'
+    jisyo_hepburn = 'hepburndict3.db'
+    jisyo_passport = 'passportdict3.db'
+    jisyo_kunrei = 'kunreidict3.db'
